Fixed incorrect forces in linear bias and ALB and update documentation
In development I tried many different force types and ALB
algorithms. To simplify all these commits, I rebased the commit
history to collapse the many commits onto a much smaller
number. During the rebase a different version's force (where I was
trying quadratic forces) was merged instead of the linear force. This
commit fixes that and adds notes on the units to the documentation.